Add tests for all reported bugs
Note that especially for the nested serializers, we don't already have
an exact end-to-end tests like this. Some unit tests come close, but
the autogenerated serializers are subtly difference from the manually
declared ones. Besides, it's nice to have a reproducible test case for
each reported but to avoid introducing regressions.