Mark fields as optional iff it has a default value
Previously, fields were marked as optional (by setting required=False)
when they had the typing.Optional qualifier. However, that qualifier is
a bit of a misnomer, as it makes the field nullable, but it can still
require a value to be provided.

Instead, only mark fields as optional when they have a default value,
and make typing.Optional only mark the field as nullable through the
allow_null argument.