Rename TypeInfo.is_optional to is_nullable
The typing.Optional qualifier is a confusing name, as a type having it
might still require a valid. Rename the field indicating this qualifier
to is_nullable to reduce the confusion.

Note that the functions in typing_utils detecting this qualifier are
still called optional, as that is a lower-level module and consistency
with the Python naming conventions is more useful there.