Elaborate why `default` field argument is not set
The choice to have updates not overwrite non-supplied fields with the
defaults is somewhat arbitrary, but I think it follows the principle of
least surprise. There's an argument to be made for overwriting, as
non-overwriting behaviour is available using partial updates anyway, but
beyond the doubtful semantics it doesn't actually gain us anything.

The main advantage of setting the `default` argument would be that it
could allow us to get rid of the `empty` sentinel and let the serializer
handle all that. However, partial updates break that dream; since as
long as we return a dataclass from `validated_data`, we need to supply
some value for non-supplied fields.

The downside of not setting the `default` argument is that it looses
documentation value when used with schema generators.