Also override validated_data when many=True is set
This requires us to create our own list serializer class, and
instantiate that class when many=True is passed.

I'm still not too happy about when and how we strip the empty sentinel
value. It seems like we can spare ourselves quite a lot of pain by doing
it earlier (or not having the value at all), but I haven't figured out
how to do that without breaking any features.